Fix: Kill server in another way
Previously we closed the server by sending the `Ctrl + C` keystroke
in its CMD window. Unfortunately this exit signal started to be
received by Deploy as well, so Deploy closed too.
We can ignore this exit signal with `SetConsoleCtrlHandler(null, true)`
but the server still doesn't stop.

Also, the server CMD window is not found anymore.
Now we find the server by its javaw process.

New solution:
Kill the java process.
Then the CMD window will prompt "Press any key to close the window"
so we also send the Enter key.

The "server-name" configuration key is now mandatory.
It is case-insensitive, and it's used to find the server Java process
(search by process name with Contains).